requirements.txt
streamlit
python-dotenv
requests
openai
beautifulsoup4
